# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Chirpy application",
        "title": "Chirpy API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/metrics": {
            "get": {
                "description": "Returns an HTML page showing the number of times Chirpy has been visited",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Get Metrics",
                "operationId": "get-metrics",
                "responses": {
                    "200": {
                        "description": "HTML content displaying the visit count",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/reset": {
            "post": {
                "description": "Delete all users in database and reset the number of times Chirpy has been visisted to 0",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Reset Metrics",
                "operationId": "post-reset",
                "responses": {
                    "200": {
                        "description": "Hits set to 0",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete users in db",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/chirps": {
            "get": {
                "description": "Retreive all chirps from database in ascendent order of created_at time",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chirp"
                ],
                "summary": "Get all chirps",
                "operationId": "get-chirps",
                "responses": {
                    "200": {
                        "description": "created chirp's information",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Chirp"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error: can not deal with data properly",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new chirp message instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chirp"
                ],
                "summary": "Create new chirp",
                "operationId": "post-create-chirp",
                "parameters": [
                    {
                        "description": "chrip message body and the author's ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.chirpParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created chirp's information",
                        "schema": {
                            "$ref": "#/definitions/models.Chirp"
                        }
                    },
                    "400": {
                        "description": "Chirp is too long, should be less than 140 chars",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error: can not deal with data properly",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/chirps/{chirpID}": {
            "get": {
                "description": "Retreive the chirp from database by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chirp"
                ],
                "summary": "Get a specific chirp",
                "operationId": "get-chirp-by-id",
                "responses": {
                    "200": {
                        "description": "Created chirp's information",
                        "schema": {
                            "$ref": "#/definitions/models.Chirp"
                        }
                    },
                    "400": {
                        "description": "Invalid chirp id provided",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Chirp not found in database",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/healthz": {
            "get": {
                "description": "Returns OK if the server is healthy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Check if the users are who they claimed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Login users",
                "operationId": "post-user-login",
                "parameters": [
                    {
                        "description": "user email and passowrd",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.userParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user's information",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "Incorrect email or password",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error: can not deal with data properly",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "description": "Create a new Chirpy user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "operationId": "post-create-user",
                "parameters": [
                    {
                        "description": "user email to get new user created",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.userParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created user's information",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "500": {
                        "description": "Internal server error occured",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/validate_chirp": {
            "post": {
                "description": "Validate a chirp's length (should be less than 14 characters) and replace profane words with ****",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "summary": "Validate a chirp",
                "operationId": "post-validation",
                "parameters": [
                    {
                        "description": "Chirp content to validate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.parameters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cleaned chirp body with profanity removed",
                        "schema": {
                            "$ref": "#/definitions/handlers.returnValues"
                        }
                    },
                    "400": {
                        "description": "Chirp is too long, should be less than 140 chars",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to decode parameters",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.chirpParams": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handlers.parameters": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "handlers.returnValues": {
            "type": "object",
            "properties": {
                "cleaned_body": {
                    "type": "string"
                }
            }
        },
        "handlers.userParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "helpers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Chirp": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}